Improved layout manager classes
