+++ src/Files.App/Data/Items/DetailsLayoutColumnItem.cs
@@ -3,7 +3,7 @@
 
 using Microsoft.UI.Xaml;
 
-namespace Files.App.Data.Models
+namespace Files.App.Data.Items
 {
 	/// <summary>
 	/// Represents item for a column shown in <see cref="DetailsLayoutPage"/>.
