+++ src/Files.App/Helpers/Layout/LayoutPreferencesManager.cs
@@ -3,26 +3,37 @@
 
 namespace Files.App.Helpers
 {
-	public class LayoutPreferences
+	public class LayoutPreferencesManager
 	{
+		// Dependency injections
+
 		private IUserSettingsService UserSettingsService { get; } = Ioc.Default.GetRequiredService<IUserSettingsService>();
 
-		public SortOption DirectorySortOption;
-		public SortDirection DirectorySortDirection;
-		public bool SortDirectoriesAlongsideFiles;
-		public GroupOption DirectoryGroupOption;
-		public SortDirection DirectoryGroupDirection;
-		public GroupByDateUnit DirectoryGroupByDateUnit;
-		public FolderLayoutModes LayoutMode;
-		public int GridViewSize;
-		public bool IsAdaptiveLayoutOverridden;
+		// Fields
 
 		public ColumnsViewModel ColumnsViewModel;
 
-		[LiteDB.BsonIgnore]
-		public static LayoutPreferences DefaultLayoutPreferences => new LayoutPreferences();
+		public bool SortDirectoriesAlongsideFiles;
+		public bool IsAdaptiveLayoutOverridden;
+		public int GridViewSize;
 
-		public LayoutPreferences()
+		public FolderLayoutModes LayoutMode;
+
+		public SortOption DirectorySortOption;
+		public SortDirection DirectorySortDirection;
+		public SortDirection DirectoryGroupDirection;
+
+		public GroupOption DirectoryGroupOption;
+		public GroupByDateUnit DirectoryGroupByDateUnit;
+
+		// Properties
+
+		[LiteDB.BsonIgnore]
+		public static LayoutPreferencesManager DefaultLayoutPreferences => new();
+
+		// Constructor
+
+		public LayoutPreferencesManager()
 		{
 			var defaultLayout = UserSettingsService.FoldersSettingsService.DefaultLayoutMode;
 
@@ -69,6 +80,8 @@ namespace Files.App.Helpers
 			ColumnsViewModel.StatusColumn.UserLengthPixels = UserSettingsService.FoldersSettingsService.SyncStatusColumnWidth;
 		}
 
+		// Overridden methods
+
 		public override bool Equals(object? obj)
 		{
 			if (obj is null)
@@ -77,7 +90,7 @@ namespace Files.App.Helpers
 			if (obj == this)
 				return true;
 
-			if (obj is LayoutPreferences prefs)
+			if (obj is LayoutPreferencesManager prefs)
 			{
 				return (
 					prefs.LayoutMode == LayoutMode &&
