+++ src/Files.App/ViewModels/Settings/FoldersViewModel.cs
@@ -454,7 +454,7 @@ namespace Files.App.ViewModels.Settings
 		public void ResetLayoutPreferences()
 		{
 			// Is this proper practice?
-			var dbInstance = FolderSettingsViewModel.GetDbInstance();
+			var dbInstance = LayoutSettingsManager.GetDbInstance();
 
 			dbInstance.ResetAll();
 		}
