+++ src/Files.App/Views/HomePage.xaml.cs
@@ -22,7 +22,7 @@ namespace Files.App.Views
 
 		private IShellPage AppInstance { get; set; } = null!;
 
-		public FolderSettingsViewModel FolderSettings
+		public LayoutSettingsManager FolderSettings
 			=> AppInstance?.InstanceViewModel.FolderSettings!;
 
 		private QuickAccessWidget? quickAccessWidget;
