+++ src/Files.App/Views/Shells/ModernShellPage.xaml.cs
@@ -83,7 +83,7 @@ namespace Files.App.Views.Shells
 			if (FilesystemViewModel is null)
 				return;
 
-			FolderSettingsViewModel.SetLayoutPreferencesForPath(FilesystemViewModel.WorkingDirectory, e.LayoutPreference);
+			LayoutSettingsManager.SetLayoutPreferencesForPath(FilesystemViewModel.WorkingDirectory, e.LayoutPreference);
 			if (e.IsAdaptiveLayoutUpdateRequired)
 				AdaptiveLayoutHelpers.ApplyAdaptativeLayout(InstanceViewModel.FolderSettings, FilesystemViewModel.WorkingDirectory, FilesystemViewModel.FilesAndFolders);
 		}
@@ -321,7 +321,7 @@ namespace Files.App.Views.Shells
 					string.IsNullOrEmpty(navArg) ||
 					!navArg.StartsWith("tag:"))) // Return if already selected
 				{
-					if (InstanceViewModel?.FolderSettings is FolderSettingsViewModel fsModel)
+					if (InstanceViewModel?.FolderSettings is LayoutSettingsManager fsModel)
 						fsModel.IsLayoutModeChanging = false;
 
 					return;
